client-service/requirements.txt
flask==3.0.3
werkzeug==3.0.3
markupsafe==2.1.5
Flask-SQLAlchemy==3.1.1
PyJWT==2.4.0
python-dotenv==1.0.1
